+++ joulescope_ui/widgets/signal_record/signal_record_config_widget.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Jetperch LLC
+# Copyright 2023-2024 Jetperch LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,18 +13,31 @@
 # limitations under the License.
 
 from PySide6 import QtWidgets, QtGui, QtCore
-from joulescope_ui import CAPABILITIES, register, pubsub_singleton, N_, get_topic_name, time64
+from joulescope_ui import CAPABILITIES, register, pubsub_singleton, N_, get_topic_name
+from joulescope_ui.filename_formatter import filename_tooltip, filename_formatter
 from joulescope_ui.styles import styled_widget
 import logging
 import os
+
+
+_FILENAME_DEFAULT = '{timestamp}.jls'
+_FILENAME_TOPIC = 'registry/SignalRecordConfigWidget/settings/filename'
 
 
 @register
 @styled_widget(N_('SignalRecordConfig'))
 class SignalRecordConfigWidget(QtWidgets.QWidget):
     CAPABILITIES = []  # no widget, since not directly instantiable
+    SETTINGS = {
+        'filename': {
+            'dtype': 'str',
+            'brief': N_('The filename with optional replacements.'),
+            'default': _FILENAME_DEFAULT,
+        },
+    }
 
     def __init__(self, parent=None):
+        self.SETTINGS = {}
         self._menu = None
         self._dialog = None
         self._row = 0
@@ -33,40 +46,52 @@ class SignalRecordConfigWidget(QtWidgets.QWidget):
         self.setSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
         self._layout = QtWidgets.QGridLayout(self)
 
+        filename = pubsub_singleton.query(_FILENAME_TOPIC)
         self._filename_label = QtWidgets.QLabel(N_('Filename'), self)
         self._filename = QtWidgets.QLineEdit(self)
-        self._filename.setText(time64.filename('.jls'))
+        self._filename.setText(filename)
+        self._filename.setToolTip(filename_tooltip())
+        self._file_reset = QtWidgets.QPushButton()
+        self._file_reset.pressed.connect(self._on_file_reset)
+        icon = self._file_reset.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_DialogDiscardButton)
+        self._file_reset.setIcon(icon)
         self._file_sel = QtWidgets.QPushButton(self)
         self._file_sel.pressed.connect(self._on_file_button)
         icon = self._file_sel.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_FileIcon)
         self._file_sel.setIcon(icon)
         self._layout.addWidget(self._filename_label, self._row, 0, 1, 1)
         self._layout.addWidget(self._filename, self._row, 1, 1, 1)
-        self._layout.addWidget(self._file_sel, self._row, 2, 1, 1)
+        self._layout.addWidget(self._file_reset, self._row, 2, 1, 1)
+        self._layout.addWidget(self._file_sel, self._row, 3, 1, 1)
         self._row += 1
 
         self._location_label = QtWidgets.QLabel(N_('Directory'), self)
         self._location = QtWidgets.QLineEdit(self)
         self._location.setText(pubsub_singleton.query('registry/paths/settings/path'))
+        self._location_reset = QtWidgets.QPushButton()
+        self._location_reset.pressed.connect(self._on_location_reset)
+        icon = self._location_reset.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_DialogDiscardButton)
+        self._location_reset.setIcon(icon)
         self._location_sel = QtWidgets.QPushButton(self)
         self._location_sel.pressed.connect(self._on_location_button)
         icon = self._location_sel.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_DirIcon)
         self._location_sel.setIcon(icon)
         self._layout.addWidget(self._location_label, self._row, 0, 1, 1)
         self._layout.addWidget(self._location, self._row, 1, 1, 1)
-        self._layout.addWidget(self._location_sel, self._row, 2, 1, 1)
+        self._layout.addWidget(self._location_reset, self._row, 2, 1, 1)
+        self._layout.addWidget(self._location_sel, self._row, 3, 1, 1)
         self._row += 1
 
         self._source_widgets = {}
         self._signals_to_record_label = QtWidgets.QLabel(N_('Signals to record'), self)
-        self._layout.addWidget(self._signals_to_record_label, self._row, 0, 1, 3)
+        self._layout.addWidget(self._signals_to_record_label, self._row, 0, 1, 4)
         self._row += 1
         self._sources_add()
 
         self._notes_label = QtWidgets.QLabel(N_('Notes'), self)
-        self._layout.addWidget(self._notes_label, self._row, 0, 1, 3)
+        self._layout.addWidget(self._notes_label, self._row, 0, 1, 4)
         self._notes = QtWidgets.QTextEdit(self)
-        self._layout.addWidget(self._notes, self._row + 4, 0, 4, 3)
+        self._layout.addWidget(self._notes, self._row + 4, 0, 4, 4)
         self._row += 5
 
     def _sources_add(self):
@@ -104,7 +129,9 @@ class SignalRecordConfigWidget(QtWidgets.QWidget):
             self._source_widgets[source] = (signal_widgets, signal_widget, signal_layout, label, spacer)
 
     def config(self):
-        path = os.path.join(self._location.text(), self._filename.text())
+        filename = self._filename.text()
+        pubsub_singleton.publish(_FILENAME_TOPIC, filename)
+        path = os.path.join(self._location.text(), filename)
         signals = []
         for source, values in self._source_widgets.items():
             topic = get_topic_name(source)
@@ -117,6 +144,11 @@ class SignalRecordConfigWidget(QtWidgets.QWidget):
             'notes': self._notes.toPlainText(),
         }
 
+    @QtCore.Slot()
+    def _on_file_reset(self):
+        self._filename.setText(_FILENAME_DEFAULT)
+
+    @QtCore.Slot()
     def _on_file_button(self):
         path = os.path.join(self._location.text(), self._filename.text())
         filter_ = 'Joulescope Data (*.jls)'
@@ -128,6 +160,7 @@ class SignalRecordConfigWidget(QtWidgets.QWidget):
         self._dialog.open()
         self._dialog.finished.connect(self._on_file_dialog_finished)
 
+    @QtCore.Slot(int)
     def _on_file_dialog_finished(self, result):
         if result == QtWidgets.QDialog.DialogCode.Accepted:
             files = self._dialog.selectedFiles()
@@ -139,6 +172,11 @@ class SignalRecordConfigWidget(QtWidgets.QWidget):
         self._dialog.close()
         self._dialog = None
 
+    @QtCore.Slot()
+    def _on_location_reset(self):
+        self._location.setText(pubsub_singleton.query('registry/paths/settings/path'))
+
+    @QtCore.Slot()
     def _on_location_button(self):
         path = self._location.text()
         self._dialog = QtWidgets.QFileDialog(self.parent(), N_('Select save location'), path)
@@ -147,6 +185,7 @@ class SignalRecordConfigWidget(QtWidgets.QWidget):
         self._dialog.open()
         self._dialog.finished.connect(self._on_location_dialog_finished)
 
+    @QtCore.Slot(int)
     def _on_location_dialog_finished(self, result):
         if result == QtWidgets.QDialog.DialogCode.Accepted:
             files = self._dialog.selectedFiles()
@@ -162,11 +201,13 @@ class SignalRecordConfigWidget(QtWidgets.QWidget):
 
 
 class SignalRecordConfigDialog(QtWidgets.QDialog):
+    count = 0
 
-    def __init__(self):
+    def __init__(self, parent=None, is_action_show=None):
         self._log = logging.getLogger(f'{__name__}.dialog')
         self._log.info('start')
-        parent = pubsub_singleton.query('registry/ui/instance')
+        if parent is None:
+            parent = pubsub_singleton.query('registry/ui/instance', default=None)
         super().__init__(parent=parent)
         self.setObjectName("signal_record_config_dialog")
         self.setAttribute(QtCore.Qt.WA_DeleteOnClose)
@@ -184,7 +225,8 @@ class SignalRecordConfigDialog(QtWidgets.QDialog):
         self._buttons.accepted.connect(self.accept)
         self._buttons.rejected.connect(self.reject)
         self._layout.addWidget(self._buttons)
-        self.finished.connect(self._on_finished)
+        if bool(is_action_show):
+            self.finished.connect(self._on_is_action_show_finished)
 
         self.resize(600, 400)
         self.setWindowTitle(N_('Configure signal recording'))
@@ -192,12 +234,13 @@ class SignalRecordConfigDialog(QtWidgets.QDialog):
         self.open()
 
     @QtCore.Slot(int)
-    def _on_finished(self, value):
+    def _on_is_action_show_finished(self, value):
         self._log.info('finished: %d', value)
-
         if value == QtWidgets.QDialog.DialogCode.Accepted:
             self._log.info('finished: accept - start recording')
             config = self._w.config()
+            config['path'] = filename_formatter(config['path'], SignalRecordConfigDialog.count)
+            SignalRecordConfigDialog.count += 1
             pubsub_singleton.publish('registry/SignalRecord/actions/!start', config)
         else:
             self._log.info('finished: reject - abort recording')
@@ -206,4 +249,4 @@ class SignalRecordConfigDialog(QtWidgets.QDialog):
 
     @staticmethod
     def on_cls_action_show():
-        SignalRecordConfigDialog()
+        SignalRecordConfigDialog(is_action_show=True)