+++ CHANGELOG.md
@@ -7,12 +7,14 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.1.0
 
-2024 Mar 21 [in progress]
+2024 Mar 23 [in progress]
 
 * Added plugin framework with live code reload  #14
 * Restructured code to support human language translation.
 * Added AI translations: ar, de, el, es, fr, ja, ko, zh.
 * Added Intel graphics dialog to display sequence  #245
+* Modified QComboBox style.
+* Added filename replacement variables.
 * Added Trigger widget.  [work in progress]