+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ---
 
-## 1.1.0
+## 1.1.1
 
 2024 Apr 3